Refactor: 폴더명 및 폴더 구조 변경 ex) Feat: Add OAuth2
