+++ src/pages/CreateFormPage.tsx
@@ -1,6 +1,5 @@
-import { useContext } from 'react';
 import styled from 'styled-components';
-import FieldList from '../components/Field/FieldList';
+import FieldList from '../components/CreateForm/FieldList';
 import CreateFormProvider from 'context/CreateFormContext';
 import CreateFormButtonWrap from 'components/CreateFormButtonWrap';
